add jmh packages
- jmh folder for testing performance
- modified build.gradle
- sample for perf-test
