+++ build.gradle
@@ -7,6 +7,7 @@
 plugins {
     id 'java-library'
     id 'maven-publish'
+    id "me.champeau.jmh" version "0.7.2"
 }
 
 repositories {
@@ -62,4 +63,11 @@ tasks.withType(JavaExec).configureEach {
 
 tasks.withType(Javadoc).configureEach {
     options.encoding = 'UTF-8'
+}
+
+jmh {
+    warmupIterations = 2
+    iterations = 5
+    fork = 1
+    duplicateClassesStrategy = DuplicatesStrategy.WARN
 }
