+++ gradle.properties
@@ -1,0 +1,3 @@
+org.gradle.jvmargs=-Xmx8g -Dfile.encoding=UTF-8
+org.gradle.parallel=true
+org.gradle.caching=true
